chore: deleted unused files and restructured
