+++ .gitignore
@@ -18,6 +18,7 @@
 .env.development.local
 .env.test.local
 .env.production.local
+react-app-env.d
 
 npm-debug.log*
 yarn-debug.log*
